refactor: Move out style code on container
